+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - If installed, use playwright to login with external browser. Please 
   [read here](https://playwright.dev/python/docs/intro) how to install playwright.
   Then use `audible.Authenticator.from_login_external(COUNTRY_CODE)` for login.
+- fix login issues
 
 ## [0.6.0] - 2021-10-21