+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+-
+
+## [0.7.0] - 2021-10-25
+
 ### Bugfix
 
 - make sure activation bytes has 8 bytes, otherwise append '0' in front until 8 bytes are reached